Move 'raster' serializers
