+++ django-rgd-imagery/rgd_imagery/serializers/__init__.py
@@ -1,10 +1,13 @@
 from .base import ImageMetaSerializer, ImageSerializer, ImageSetSerializer
 from .processed import ConvertedImageSerializer, RegionImageSerializer
+from .raster import RasterMetaSerializer, RasterSerializer
 
 __all__ = [
     'ConvertedImageSerializer',
     'ImageSerializer',
     'ImageMetaSerializer',
     'ImageSetSerializer',
+    'RasterMetaSerializer',
+    'RasterSerializer',
     'RegionImageSerializer',
 ]
